Host the website via the server
Add an optional server for hosting the frontend. The REST API can be hosted at either a separate domain (as is done in production) or as a subdomain (like /api). There is a build script to pre-zip all the frontend files, which can be served zipped to clients which accept gzip.

There are four values for `SR_HOST_FRONTEND`:
+ "" (empty):  do not host the frontend. Do not redirect to FrontendOrigin at the root.
+ "redirect":  Redirect to FrontendOrigin from the root.
+ "subroute":  Host the frontend at / and /static. Host the REST API at /api.
+ "by-domain": Filter requests by the Host header, FrontendOrigin to frontend and BackendOrigin to backend. Redirect requests with other Host to FrontendOrigin.

- Add frontend server with compression support via precompressed files and ETag via hashed filenames
- Configuration options to redirect to frontend and separate hosts via domain or subpath (see above)
- Scripts to perform gzip compression and removal of compressed files
- Task to fix issues with presite on the frontend
- Bump required go version to 1.16 to use fs library

https://github.com/SnirkImmington/sr-server/pull/21