Add 'staging' Docker configuration
This tries to match a production environment while running locally.
The frontend is set to NODE_ENV=production and the backend has
SR_IS_PRODUCTION=true. The frontend is served via Shadowroller's
frontend server, so this is also the way to test that server locally.
The API is served on /api.
This will probably more closely match the configuration used by self-
hosting users, so we will keep this configuration in the codebase.
It's currently HTTP only and not cross-domain, so a future version
should use self-signed certs and exercise CORS code more.