Bump rate limit for local testing
The preveious rate limit (5 req / 10 s) was designed to allow
the developer to hit the rate limit when testing locally. This
did not take frontend access into account - the frontend hits
~5 resources when loading the main page.
This difference in traffic between the frontend and API should
be taken into account when rate limiting in the future.